app/requirements.txt
anyio==3.6.1
appdirs==1.4.4
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.11.1
black==22.3.0
bleach==5.0.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
commonmark==0.9.1
contextvars==2.4
coverage==6.4.2
cycler==0.11.0
debugpy==1.6.2
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.4
docutils==0.19
entrypoints==0.4
executing==0.8.3
fastapi==0.78.0
fastjsonschema==2.15.3
filelock==3.7.1
flake8==4.0.1
fonttools==4.34.4
fsspec==2022.5.0
funcsigs==1.0.2
h11==0.13.0
htmltools
humanize==4.2.3
idna==3.3
immutables==0.18
importlib-metadata==4.12.0
importlib-resources==5.8.0
iniconfig==1.1.1
ipykernel==6.15.1
ipyleaflet==0.17.0
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
jsonschema==4.7.2
jupyter-client==7.3.4
jupyter-core==4.11.1
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.1
keyring==23.6.0
kiwisolver==1.4.3
linkify-it-py==2.0.0
markdown-it-py==2.1.0
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mccabe==0.6.1
mdurl==0.1.1
mistune==0.8.4
mizani==0.7.4
mypy==0.961
mypy-extensions==0.4.3
nbclient==0.6.6
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
notebook==6.4.12
numpy==1.23.1
packaging==21.3
palettable==3.3.0
pandas==1.4.3
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
patsy==0.5.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pins==0.5.0
pipenv==2022.7.4
pkginfo==1.8.3
platformdirs==2.5.2
plotnine==0.9.0
pluggy==1.0.0
prometheus-client==0.14.1
prompt-toolkit==3.0.30
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.9.1
pyflakes==2.4.0
Pygments==2.12.0
pyparsing==3.0.9
pyrsistent==0.18.1
pytest==7.1.2
pytest-cov==3.0.0
pytest-mypy==0.9.1
python-dateutil==2.8.2
python-dotenv==0.20.0
python-multipart==0.0.5
pytz==2022.1
PyYAML==6.0
pyzmq==23.2.0
readme-renderer==35.0
requests==2.28.1
requests-toolbelt==0.9.1
rfc3986==2.0.0
rich==12.5.1
scikit-learn==1.1.1
scipy==1.8.1
semver==2.13.0
Send2Trash==1.8.0
setuptools-scm==7.0.5
shiny
shinywidgets
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.2.post1
stack-data==0.3.0
starlette==0.19.1
statsmodels==0.13.2
terminado==0.15.0
threadpoolctl==3.1.0
tinycss2==1.1.1
toml==0.10.2
tomli==2.0.1
tornado==6.2
traitlets==5.3.0
traittypes==0.2.1
twine==4.0.1
types-click==7.1.8
types-Flask==1.1.6
types-Jinja2==2.11.9
types-MarkupSafe==1.1.10
types-six==1.16.18
types-Werkzeug==1.0.9
typing_extensions==4.3.0
uc-micro-py==1.0.1
urllib3==1.26.10
uvicorn==0.18.2
vetiver==0.1.4
virtualenv==20.15.1
virtualenv-clone==0.5.7
wcwidth==0.2.5
webencodings==0.5.1
websockets==10.3
widgetsnbextension==3.6.1
xxhash==3.0.0
xyzservices==2022.6.0
zipp==3.8.1
--extra-index-url https://pyshiny.netlify.app/pypi/
